src/main/java/ru/egar/spring_work_accounting/employee/EmployeeController.java
package ru.egar.spring_work_accounting.employee;

import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import ru.egar.spring_work_accounting.abstraction.exceptions.ExceptionResponse;

import java.util.UUID;

@RestController
@RequestMapping("employee")
@RequiredArgsConstructor
public class EmployeeController {

    private final EmployeeAdapterService employeeAdapterService;

    @GetMapping("{id}")
    public ResponseEntity<EmployeeResponse> findEmployeeById(@PathVariable UUID id) {
        return ResponseEntity.ok(employeeAdapterService.findById(id));
    }

    @PostMapping
    public ResponseEntity<UUID> createEmployee(@RequestBody EmployeeRequest employeeRequest) {
        return new ResponseEntity<>(employeeAdapterService.save(employeeRequest), HttpStatus.CREATED);
    }

    @PatchMapping("{id}")
    public ResponseEntity<UUID> updateEmployee(@RequestBody EmployeeRequest employeeRequest, @PathVariable UUID id) {
        return ResponseEntity.ok(employeeAdapterService.update(employeeRequest, id));
    }

    @DeleteMapping("{id}")
    public ResponseEntity<HttpStatus> deleteEmployee(@PathVariable UUID id) {
        employeeAdapterService.delete(id);
        return new ResponseEntity<>(HttpStatus.OK);
    }

    @ExceptionHandler(value = EmployeeNotFoundException.class)
    public ResponseEntity<ExceptionResponse> handleEmployeeNotFoundException(EmployeeNotFoundException ex) {
        return new ResponseEntity<>(new ExceptionResponse(ex.getMessage(), System.currentTimeMillis()), HttpStatus.BAD_REQUEST);
    }

}
